nexus-rag/nexus-rag/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
endee==0.1.25
endee-model==0.1.0
sentence-transformers==3.0.1
httpx>=0.28.1
pydantic==2.7.4
python-multipart==0.0.9
aiofiles==23.2.1
numpy>=2.2.4
tiktoken==0.7.0
pypdf==4.2.0
sse-starlette==2.1.0
rich==13.7.1
